Store range attack direction when the ability starts
Testing also customizing instance parameters in code, I am thinking about this or just a list of components, or maybe a callback and configure everything before the system initializes. Not sure, in fact I have another way, a list of raw parameters to be optionally processed by each system on initialization.